Integrate dropdown with formik
